Add ethereum address generator and update ecdh key exchange
