+++ crypto-notes/README.md
@@ -1,5 +1,5 @@
 ```
 tsc && node ecdsaVerify.js
 tsc && node keyPair.js
-tsc && node ecdh.js
+tsc && node ethereumAddressGenerator.js
 ```
